Uploaded new version: Program Overflow 2.1

+++ Form1.cs
@@ -1,5 +1,5 @@
 /*
- * Program Overflow 2.0 coded by LumitoLuma.
+ * Program Overflow 2.1 coded by LumitoLuma.
  * (C) 2020, LumitoLuma.
  * Please use this tool responsibly.
  * I'm not responsible of any damage this tool can make to your PC.
@@ -17,88 +17,13 @@ namespace ProgramOverflow
         byte counterRipPc;
         byte counterRunPrograms;
         byte ischk;
+        byte ischk2;
         byte RunPrograms;
         short x;
         int i;
         int ProgressBar1Value;
 
-        public void CopyArg(int argnumber)
-        {
-            if (argnumber <= 999999)
-            {
-                if (!Directory.Exists("C:\\ProgramOverflow"))
-                {
-                    Directory.CreateDirectory("C:\\ProgramOverflow");
-                }
-                for (i = 1; i <= argnumber; i++)
-                {
-                    string destinationFile = "C:\\ProgramOverflow\\Program" + i + ".exe";
-                    var process = System.Diagnostics.Process.GetCurrentProcess();
-                    string fullPath = process.MainModule.FileName;
-                    if (!File.Exists(destinationFile))
-                    {
-                        File.Copy(fullPath, destinationFile, true);
-                    }
-                    if (RunPrograms == 1)
-                    {
-                        System.Diagnostics.Process.Start(destinationFile);
-                    }
-                }
-                MessageBox.Show("Copied " + argnumber.ToString() + " files correctly!", "Program Overflow 2.0");
-                Environment.Exit(0);
-            }
-            else
-            {
-                MessageBox.Show("Please specify a number between 1 and 999999. 0 yields error", "Program Overflow 2.0");
-                Environment.Exit(0);
-            }
-        }
-
-        public static void Clean()
-        {
-            if (Directory.Exists("C:\\ProgramOverflow"))
-            {
-                DirectoryInfo podir = new DirectoryInfo("C:\\ProgramOverflow");
-
-                int fc = 0;
-
-                foreach (FileInfo file in podir.EnumerateFiles())
-                {
-                    file.Delete(); //Deletes all the files in C:\ProgramOverflow directory
-                    fc++;
-                }
-                foreach (DirectoryInfo dir in podir.EnumerateDirectories())
-                {
-                    dir.Delete(true); //Deletes all the directories in C:\ProgramOverflow directory
-                }
-                podir.Delete(); //Deletes C:\ProgramOverflow directory
-                if (fc == 1) //Checks if there was only a file in C:\ProgramOverflow directory
-                {
-                    MessageBox.Show("1 file successfully cleaned!", "Program Overflow 2.0");
-                }
-                else
-                {
-                    MessageBox.Show(fc + " files successfully cleaned!", "Program Overflow 2.0");
-                }
-                Environment.Exit(0); //Exits the program
-            }
-            else
-            {
-                MessageBox.Show("No need to clean", "Program Overflow 2.0");
-                Environment.Exit(0);
-            }
-        }
-
-        public Form1()
-        {
-            InitializeComponent();
-            counterRipPc = 0;
-            counterRunPrograms = 0;
-            ischk = 0;
-            RunPrograms = 0;
-        }
-
-        private void Button1_Click(object sender, EventArgs e)
+        public void CopyFiles()
         {
             button1.Enabled = false;
             button2.Enabled = false;
@@ -135,7 +60,7 @@ namespace ProgramOverflow
                 catch
                 {
                     richTextBox1.AppendText(" Error!\n-- Error while deleting files --");
-                }                
+                }
             }
             progressBar1.Value = progressBar1.Maximum;
             button1.Enabled = true;
@@ -143,27 +68,56 @@ namespace ProgramOverflow
             richTextBox1.AppendText("-- " + (i - 1) + " Files successfully copied! --");
         }
 
-        private void Button2_Click(object sender, EventArgs e)
+        public void CopyArg(int argnumber)
         {
-            button1.Enabled = false;
-            button2.Enabled = false;
-            richTextBox1.Clear();
-            progressBar1.Value = 0;
+            if (argnumber <= 999999)
+            {
+                if (!Directory.Exists("C:\\ProgramOverflow"))
+                {
+                    Directory.CreateDirectory("C:\\ProgramOverflow");
+                }
+                for (i = 1; i <= argnumber; i++)
+                {
+                    string destinationFile = "C:\\ProgramOverflow\\Program" + i + ".exe";
+                    var process = System.Diagnostics.Process.GetCurrentProcess();
+                    string fullPath = process.MainModule.FileName;
+                    if (!File.Exists(destinationFile))
+                    {
+                        File.Copy(fullPath, destinationFile, true);
+                    }
+                    if (RunPrograms == 1)
+                    {
+                        System.Diagnostics.Process.Start(destinationFile);
+                    }
+                }
+                MessageBox.Show("Copied " + argnumber.ToString() + " files correctly!", "Program Overflow 2.1");
+                Environment.Exit(0);
+            }
+            else
+            {
+                MessageBox.Show("Please specify a number between 1 and 999999. 0 yields error", "Program Overflow 2.1");
+                Environment.Exit(0);
+            }
+        }
+
+        public static void Clean()
+        {
             if (Directory.Exists("C:\\ProgramOverflow"))
             {
-                DirectoryInfo podir = new DirectoryInfo("C:\\ProgramOverflow");
-                int fc = 0;
-                int nfiles = podir.GetFiles().Length;
-                progressBar1.Maximum = nfiles * 100;
-                ProgressBar1Value = progressBar1.Maximum / nfiles;
-                try
+                string msg = "This will delete the whole C:\\ProgramOverflow directory.\nAre you sure that you want to continue?";
+                string titlemsg = "Program Overflow 2.1";
+                MessageBoxButtons buttons = MessageBoxButtons.YesNo;
+                MessageBoxDefaultButton dbutton = MessageBoxDefaultButton.Button2;
+                DialogResult result = MessageBox.Show(msg, titlemsg, buttons, MessageBoxIcon.Warning, dbutton);
+                if (result == DialogResult.Yes)
                 {
+                    DirectoryInfo podir = new DirectoryInfo("C:\\ProgramOverflow");
+
+                    int fc = 0;
+
                     foreach (FileInfo file in podir.EnumerateFiles())
                     {
-                        progressBar1.Value += ProgressBar1Value;
-                        richTextBox1.AppendText("Deleting " + file + "...");
                         file.Delete(); //Deletes all the files in C:\ProgramOverflow directory
-                        richTextBox1.AppendText(" Done!\n");
                         fc++;
                     }
                     foreach (DirectoryInfo dir in podir.EnumerateDirectories())
@@ -173,25 +127,102 @@ namespace ProgramOverflow
                     podir.Delete(); //Deletes C:\ProgramOverflow directory
                     if (fc == 1) //Checks if there was only a file in C:\ProgramOverflow directory
                     {
-                        richTextBox1.AppendText("-- 1 file successfully cleaned! -- ");
+                        MessageBox.Show("1 file successfully cleaned!", "Program Overflow 2.1");
                     }
                     else
                     {
-                        richTextBox1.AppendText("-- " + fc + " files successfully cleaned! --");
+                        MessageBox.Show(fc + " files successfully cleaned!", "Program Overflow 2.1");
                     }
                 }
-                catch
+                Environment.Exit(0); //Exits the program
+            }
+            else
+            {
+                MessageBox.Show("No need to clean", "Program Overflow 2.1");
+                Environment.Exit(0);
+            }
+        }
+
+        public Form1()
+        {
+            InitializeComponent();
+            counterRipPc = 0;
+            counterRunPrograms = 0;
+            ischk = 0;
+            ischk2 = 0;
+            RunPrograms = 0;
+        }
+
+        private void Button1_Click(object sender, EventArgs e)
+        {
+            CopyFiles();
+        }
+
+        private void Button2_Click(object sender, EventArgs e)
+        {
+            int cno = 0;
+            button1.Enabled = false;
+            button2.Enabled = false;
+            richTextBox1.Clear();
+            progressBar1.Value = 0;
+            if (Directory.Exists("C:\\ProgramOverflow"))
+            {
+                string msg = "This will delete the whole C:\\ProgramOverflow directory.\nAre you sure that you want to continue?";
+                string titlemsg = "Program Overflow 2.1";
+                MessageBoxButtons buttons = MessageBoxButtons.YesNo;
+                MessageBoxDefaultButton dbutton = MessageBoxDefaultButton.Button2;
+                DialogResult result = MessageBox.Show(msg, titlemsg, buttons, MessageBoxIcon.Warning, dbutton);
+                if (result == DialogResult.Yes)
                 {
-                    richTextBox1.AppendText(" Error!\n-- Error while deleting files --");
+                    DirectoryInfo podir = new DirectoryInfo("C:\\ProgramOverflow");
+                    int fc = 0;
+                    int nfiles = podir.GetFiles().Length;
+                    progressBar1.Maximum = nfiles * 100;
+                    ProgressBar1Value = progressBar1.Maximum / nfiles;
+                    try
+                    {
+                        foreach (FileInfo file in podir.EnumerateFiles())
+                        {
+                            progressBar1.Value += ProgressBar1Value;
+                            richTextBox1.AppendText("Deleting " + file + "...");
+                            file.Delete(); //Deletes all the files in C:\ProgramOverflow directory
+                            richTextBox1.AppendText(" Done!\n");
+                            fc++;
+                        }
+                        foreach (DirectoryInfo dir in podir.EnumerateDirectories())
+                        {
+                            dir.Delete(true); //Deletes all the directories in C:\ProgramOverflow directory
+                        }
+                        podir.Delete(); //Deletes C:\ProgramOverflow directory
+                        if (fc == 1) //Checks if there was only a file in C:\ProgramOverflow directory
+                        {
+                            richTextBox1.AppendText("-- 1 file successfully cleaned! -- ");
+                        }
+                        else
+                        {
+                            richTextBox1.AppendText("-- " + fc + " files successfully cleaned! --");
+                        }
+                    }
+                    catch
+                    {
+                        richTextBox1.AppendText(" Error!\n-- Error while deleting files --");
+                    }
+                }
+                else
+                {
+                    cno = 1;
                 }
             }
             else
             {
                 richTextBox1.AppendText("-- No need to clean --");
             }
+            if(cno == 0)
+            {
+                progressBar1.Value = progressBar1.Maximum;
+            }
             button1.Enabled = true;
             button2.Enabled = true;
-            progressBar1.Value = progressBar1.Maximum;
         }
 
         private void RadioButton1_CheckedChanged(object sender, EventArgs e)
@@ -199,9 +230,10 @@ namespace ProgramOverflow
             if(ischk == 0)
             {
                 string msg = "Are you sure that you want to enable this feature?";
-                string titlemsg = "Program Overflow 2.0";
+                string titlemsg = "Program Overflow 2.1";
                 MessageBoxButtons buttons = MessageBoxButtons.YesNo;
-                DialogResult result = MessageBox.Show(msg, titlemsg, buttons, MessageBoxIcon.Warning);
+                MessageBoxDefaultButton dbutton = MessageBoxDefaultButton.Button2;
+                DialogResult result = MessageBox.Show(msg, titlemsg, buttons, MessageBoxIcon.Warning, dbutton);
                 if (result == DialogResult.Yes)
                 {
                     checkBox1.Enabled = true;
@@ -239,17 +271,39 @@ namespace ProgramOverflow
 
         private void CheckBox2_CheckedChanged(object sender, EventArgs e)
         {
-            if (counterRunPrograms == 0)
+            if (ischk2 == 0)
             {
-                RunPrograms = 1;
-                counterRunPrograms = 1;
-                button1.Text = "Destroy computer";
+                if (counterRunPrograms == 0)
+                {
+                    string msg = "This feature may DESTROY your PC.\nAre you highly sure that you want to enable it?";
+                    string titlemsg = "Program Overflow 2.1";
+                    MessageBoxButtons buttons = MessageBoxButtons.YesNo;
+                    MessageBoxDefaultButton dbutton = MessageBoxDefaultButton.Button2;
+                    DialogResult result = MessageBox.Show(msg, titlemsg, buttons, MessageBoxIcon.Stop, dbutton);
+                    if (result == DialogResult.Yes)
+                    {
+                        RunPrograms = 1;
+                        counterRunPrograms = 1;
+                        button1.Text = "Destroy computer";
+                        button1.ForeColor = System.Drawing.Color.Red;
+                    }
+                    else
+                    {
+                        ischk2 = 1;
+                        checkBox2.Checked = false;
+                    }
+                }
+                else
+                {
+                    RunPrograms = 0;
+                    counterRunPrograms = 0;
+                    button1.Text = "Start";
+                    button1.ForeColor = System.Drawing.Color.Black;
+                }
             }
             else
             {
-                RunPrograms = 0;
-                counterRunPrograms = 0;
-                button1.Text = "Start";
+                ischk2 = 0;
             }
         }
 
@@ -421,7 +475,7 @@ namespace ProgramOverflow
                                         "/S:\n" +
                                         "  Use this flag after \"Number of copies\" or \"/clean\".\n" +
                                         "  \"/S\" flag suppress any message that the program will show, so the execution is silent.\n\n" +
-                                        "© 2020, Lumito - www.lumito.net", "Program Overflow 2.0");
+                                        "© 2020, Lumito - www.lumito.net", "Program Overflow 2.1");
 
                         Environment.Exit(0);
                     }
@@ -431,6 +485,15 @@ namespace ProgramOverflow
                     // Starts the app normally
                 }
                 
+            }
+        }
+
+        private void NumericUpDown1_KeyDown(object sender, KeyEventArgs e)
+        {
+            if (e.KeyCode == Keys.Enter)
+            {
+                CopyFiles();
+                e.SuppressKeyPress = true;
             }
         }
     }
